pyproject.toml
[project]
name = "designrush-seo-audit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "polars>=1.33.1",
    "requests>=2.32.3",
    "PyYAML>=6.0.2",
    "python-dotenv>=1.0.1",
]
